requirements.txt
Flask
python-dotenv
currencyapicom
